+++ src/config.ts
@@ -1,3 +1,5 @@
+import { WebhookTriggers } from './models/webhook';
+
 export let apiServer: string | null = null;
 export function setApiServer(newApiServer: string | null) {
   apiServer = newApiServer;
@@ -28,3 +30,29 @@ export type AuthenticateUrlConfig = {
   scopes?: string[];
   responseType?: ResponseType;
 };
+
+export enum Region {
+  Us = 'us',
+  Ireland = 'ireland',
+}
+
+export const DEFAULT_REGION = Region.Us;
+
+export const regionConfig = {
+  [Region.Us]: {
+    nylasAPIUrl: 'https://api.nylas.com',
+    dashboardApiUrl: 'https://dashboard-api.nylas.com',
+    callbackDomain: 'cb.nylas.com',
+    websocketDomain: 'tunnel.nylas.com',
+    telemetryApiUrl: 'https://cli.nylas.com',
+  },
+  [Region.Ireland]: {
+    nylasAPIUrl: 'https://ireland.api.nylas.com',
+    dashboardApiUrl: 'https://ireland.dashboard.nylas.com',
+    callbackDomain: 'cb.nylas.com',
+    websocketDomain: 'tunnel.nylas.com',
+    telemetryApiUrl: 'https://cli.nylas.com',
+  },
+};
+
+export const DEFAULT_WEBHOOK_TRIGGERS = Object.values(WebhookTriggers);