+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-### Unreleased
+### 6.8.0 / 2023-02-03
 * Local webhook testing support
 * Add `hideParticipants` field to `Event`